refactor(travel): extract common 4-4 travel logic into base class
Move duplicated travel routing, tick handling, and 4-4 map management code from SmartX1Travel and SmartBlTravel into the new abstract FourFourRouter base behavior. This reduces code duplication and simplifies future maintenance of both travel behaviors.
